Projeto Finalizado

+++ Models/Produto.cs
@@ -1,3 +1,5 @@
+using System.ComponentModel.DataAnnotations.Schema;
+
 namespace ProjetoMercadinho.Models
 {
     public class Produto
@@ -6,8 +8,8 @@ namespace ProjetoMercadinho.Models
         public string Nome { get; set; }
         public Categoria Categoria { get; set; }
         public Fornecedor Fornecedor { get; set; }
-        public float PrecoDeCusto { get; set; }
-        public float PrecoDeVenda { get; set; }
+        public decimal PrecoDeCusto { get; set; }      
+        public decimal PrecoDeVenda { get; set; }
         public int Medicao { get; set; }
         public bool Status { get; set; }
     }
